requirements.txt
Flask==2.0.1
pytesseract==0.3.8
Pillow==8.2.0
opencv-python==4.5.2.52
python-dotenv==0.17.1
